+++ example/settings.py
@@ -42,6 +42,12 @@ STATIC_URL = "/static/"
 
 TEMPLATES = [
     {
+        "NAME": "jinja2",
+        "BACKEND": "django.template.backends.jinja2.Jinja2",
+        "APP_DIRS": True,
+        "DIRS": [os.path.join(BASE_DIR, "example", "templates", "jinja2")],
+    },
+    {
         "BACKEND": "django.template.backends.django.DjangoTemplates",
         "APP_DIRS": True,
         "DIRS": [os.path.join(BASE_DIR, "example", "templates")],
@@ -54,7 +60,7 @@ TEMPLATES = [
                 "django.contrib.messages.context_processors.messages",
             ],
         },
-    }
+    },
 ]
 
 USE_TZ = True