+++ example/views.py
@@ -1,4 +1,5 @@
 from django.http import JsonResponse
+from django.shortcuts import render
 
 
 def increment(request):
@@ -8,3 +9,7 @@ def increment(request):
         value = 1
     request.session["value"] = value
     return JsonResponse({"value": value})
+
+
+def jinja2_view(request):
+    return render(request, "index.jinja", {"foo": "bar"}, using="jinja2")